requirements.txt
sqlparse==0.5.1
nltk==3.9.1
openai==1.51.2
fire==0.7.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
datasets==3.0.1
jsonlines==4.0.0
tabulate==0.9.0
retry==0.9.2
func-timeout==4.3.5
psycopg2-binary==2.9.9
pymysql
google-generativeai==0.8.3
anthropic==0.36.1
